requirements/base.txt
Django
dj-rest-auth
django-allauth
django-rest-auth
oauthlib
django-axes
django-admin-interface
django-environ
django-cors-headers
django-axes
django-filter
django-log-viewer
django-flat-responsive
django-flat-theme
psycopg2-binary
django-colorfield
drf-writable-nested
drf-yasg
djangorestframework
djangorestframework-api-key
djangorestframework-simplejwt
django-anymail
wheel